backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.5.1
bcrypt==4.3.0
boto3==1.37.15
botocore==1.37.15
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.115.11
google-api-core==2.24.2
google-api-python-client==2.115.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jmespath==1.0.1
proto-plus==1.26.1
protobuf==6.30.1
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.3
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.4.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9
s3transfer==0.11.4
sendgrid==6.11.0
six==1.17.0
sniffio==1.3.1
starkbank-ecdsa==2.2.0
starlette==0.46.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
